+++ src/goa.rs
@@ -62,6 +62,7 @@ impl GoaMailAccount {
             empty_junk_days: 0,
             empty_trash_days: 0,
             pgp_key: None,
+            in_unified: true,
             name: if self.name.trim().is_empty() {
                 self.email.clone()
             } else {