+++ src/worker.rs
@@ -11344,6 +11344,7 @@ pub(super) fn sample_account() -> AccountConfig {
         empty_junk_days: 0,
         empty_trash_days: 0,
         pgp_key: None,
+        in_unified: true,
         push: None,
         name: String::new(),
         email: "me@example.com".into(),